Unify the generate/create mirrors for table, form and security duty/role
createD365File.ts and generateD365Xml.ts each expose an XmlTemplateGenerator
that is supposed to be a mirror. Three had drifted into throwing content
away:

  security duty/role  hard-coded <Privileges /> and <Duties /> regardless
                      of what the caller passed. The create path was fixed
                      after the "security chain silent empty write"
                      incident; the mirror was not.
  table               emitted a bare <Fields />, dropping every field, and
                      hand-rolled its property block in a different order —
                      and AxTable drops a misordered property silently.
  form                its parameter was named `_properties`. It ignored the
                      design pattern, data source, caption and controls
                      outright and returned a fixed empty shell.

The documented hybrid flow is generate → create(xmlContent), so that
mirror wrote empty, non-functional objects to disk. They all build
perfectly clean, which is why this went unnoticed.

Extracts the create-side (correct) implementations into shared builders —
tableXml.ts, formXml.ts, securityDutyRoleXml.ts — following the existing
securityPrivilegeXml.ts / queryViewXml.ts convention, and points both
halves at them. The duplicated helpers (fieldTypeToAxType,
normalizeNameList, securityRefContainer, refContainer) collapse into one
copy each.

The parity test asserts both that the two agree AND that the agreed
output carries the content — two empty shells would agree too. 5 of its
6 cases fail against the pre-fix mirrors.
